+++ src/main/java/open/zzj/demo/demo/dto/QuestionDto.java
@@ -7,13 +7,13 @@ import open.zzj.demo.demo.model.User;
 @Data
 public class QuestionDto {
 
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creatorId;
+    private Long creatorId;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
